updated bootstrap script

+++ script/bootstrap.sh
@@ -15,11 +15,16 @@ gsettings set org.cinnamon.settings-daemon.plugins.power sleep-display-battery 0
 echo "Installing packages"
 sudo apt install -y bat colordiff tmux vim ccze terminator
 
-echo "Installing Hack Nerd Font"
-wget -P /tmp https://github.com/ryanoasis/nerd-fonts/releases/download/v3.3.0/Hack.zip 2> /dev/null
-sudo unzip /tmp/Hack.zip -d /usr/local/share/fonts/Hack
-rm /tmp/Hack.zip
-fc-cache -fv
+# Check if the folder exists
+if [ -d "/usr/local/share/fonts/Hack/" ]; then
+    echo "Hack Nerd font already installed."
+else
+    echo "Installing Hack Nerd font"
+    wget -P /tmp https://github.com/ryanoasis/nerd-fonts/releases/download/v3.3.0/Hack.zip 2> /dev/null
+    sudo unzip /tmp/Hack.zip -d /usr/local/share/fonts/Hack
+    rm /tmp/Hack.zip
+    fc-cache -fv > /dev/null
+fi
 
 echo "Checking if docker is installed"
 if which docker > /dev/null 2>&1; then
